+++ src/com/techjar/jfos2/client/gui/GUIBackground.java
@@ -40,7 +40,7 @@ public class GUIBackground extends GUI {
     }
 
     @Override
-    public void update(double delta) {
+    public void update(float delta) {
     }
 
     @Override
